fix: refacor code and group files.
